优化多重缓存逻辑及过期时间，新增熔断限流降级措施，新增请求强制设置超时时间

+++ repository/shortUrl.go
@@ -2,6 +2,7 @@ package repository
 
 import (
 	"context"
+	"errors"
 	"short_url/repository/cache"
 	"short_url/repository/dao"
 	"time"
@@ -48,15 +49,29 @@ func (c *CachedShortUrlRepository) GetOriginUrlByShortUrl(ctx context.Context, s
 	now := time.Now().Unix()
 
 	// 先查本地缓存，若本地缓存存在直接返回
-	if val, ok := c.lru.Get(shortUrl); ok {
+	val, ok := c.lru.Get(shortUrl)
+	if ok {
 		if item, ok := val.(lruItem); ok && item.expiredAt >= now {
 			return item.originUrl, nil
 		}
 	}
 
-	// 再查redis，若redis存在直接返回
+	// 再查 redis，若 redis 存在直接返回
 	if originUrl, err := c.cache.Get(ctx, shortUrl); err == nil {
+		go func() {
+			if err := c.cache.Refresh(ctx, shortUrl); err != nil {
+				c.l.Error("failed to refresh redis cache",
+					logger.Error(err),
+					logger.String("short_url", shortUrl),
+				)
+			}
+		}()
 		return originUrl, nil
+	}
+
+	if ctx.Value("break").(bool) {
+		// 熔断模式下，直接返回本地缓存的值
+		return val.(lruItem).originUrl, nil
 	}
 
 	// 最后查数据库
@@ -69,7 +84,7 @@ func (c *CachedShortUrlRepository) GetOriginUrlByShortUrl(ctx context.Context, s
 	go func() {
 		defer cancel()
 
-		// 异步更新redis缓存
+		// 异步更新 redis 缓存
 		if err = c.cache.Set(newCtx, shortUrl, su.OriginUrl); err != nil {
 			c.l.Error("failed to set redis cache",
 				logger.Error(err),
@@ -80,7 +95,7 @@ func (c *CachedShortUrlRepository) GetOriginUrlByShortUrl(ctx context.Context, s
 		// 异步更新本地 lru 缓存
 		c.lru.Add(shortUrl, lruItem{
 			originUrl: su.OriginUrl,
-			expiredAt: int64(c.lruExpiration.Seconds()),
+			expiredAt: int64(time.Now().Add(c.lruExpiration).Unix()),
 		})
 	}()
 
@@ -88,6 +103,11 @@ func (c *CachedShortUrlRepository) GetOriginUrlByShortUrl(ctx context.Context, s
 }
 
 func (c *CachedShortUrlRepository) InsertShortUrl(ctx context.Context, shortUrl, originUrl string) error {
+	if ctx.Value("downgrade").(bool) {
+		// 降级模式下，禁止新增数据
+		return errors.New("downgrade")
+	}
+
 	return c.dao.Insert(ctx, dao.ShortUrl{
 		ShortUrl:  shortUrl,
 		OriginUrl: originUrl,
@@ -96,6 +116,11 @@ func (c *CachedShortUrlRepository) InsertShortUrl(ctx context.Context, shortUrl,
 }
 
 func (c *CachedShortUrlRepository) DeleteShortUrlByShortUrl(ctx context.Context, shortUrl string) error {
+	if ctx.Value("downgrade").(bool) {
+		// 降级模式下，禁止删除数据
+		return errors.New("downgrade")
+	}
+
 	err := c.dao.DeleteByShortUrl(ctx, shortUrl)
 	if err == nil {
 		newCtx, cancel := context.WithTimeout(ctx, 1*time.Second)
@@ -117,6 +142,11 @@ func (c *CachedShortUrlRepository) DeleteShortUrlByShortUrl(ctx context.Context,
 }
 
 func (c *CachedShortUrlRepository) CleanExpired(ctx context.Context, now int64) error {
+	if ctx.Value("downgrade").(bool) {
+		// 降级模式下，禁止清理过期数据
+		return errors.New("downgrade")
+	}
+
 	deleteList, err := c.dao.DeleteExpiredList(ctx, now)
 	if err == nil {
 		newCtx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
